+++ src/config.rs
@@ -1,8 +1,8 @@
 /// Main Config of window <br>
 /// Contains logical and physical sizes  <br>
 /// Logical size is your drawing canvas size
-/// Physical size is the actual window size on screen  
-/// Defaults to 800x600 logical and physical sizes
+/// Physical size is the actual window size on screen <br>
+/// **Defaults to 1280x800 physical & 640&400 logical sizes**
 pub struct Config {
     /// Window Title
     pub title: String,
@@ -16,6 +16,9 @@ pub struct Config {
     // Initial window size
     pub window_width: u32,
     pub window_height: u32,
+
+    /// Fps
+    pub target_fps: Option<u32>
 }
 pub struct ConfigBuilder {
     /// Window Title
@@ -28,11 +31,17 @@ pub struct ConfigBuilder {
     // Initial window size
     window_width: Option<u32>,
     window_height: Option<u32>,
+    /// Fps
+    target_fps: Option<u32>
 }
 impl ConfigBuilder {
     /// Set the window title
     pub fn with_title<S: Into<String>>(mut self, title: S) -> Self {
         self.title = Some(title.into());
+        self
+    }
+    pub fn with_target_fps(mut self, fps: u32) -> Self{
+        self.target_fps = Some(fps);
         self
     }
     /// Set the window size (physical size)
@@ -51,15 +60,17 @@ impl ConfigBuilder {
         self.resizeable = Some(resizeable);
         self
     }
+    
     /// Build the Config
     pub fn build(self) -> Config {
         Config {
             title: self.title.unwrap_or("Graphiclity Window".to_string()),
             resizeable: self.resizeable.unwrap_or(true),
-            logical_width: self.logical_width.unwrap_or(800),
-            logical_height: self.logical_height.unwrap_or(600),
-            window_width: self.window_width.unwrap_or(800),
-            window_height: self.window_height.unwrap_or(600),
+            logical_width: self.logical_width.unwrap_or(640),
+            logical_height: self.logical_height.unwrap_or(400),
+            window_width: self.window_width.unwrap_or(1280),
+            window_height: self.window_height.unwrap_or(800),
+            target_fps: self.target_fps,
         }
     }
 }
@@ -73,6 +84,7 @@ impl Config {
             window_width: None,
             resizeable: None,
             title: None,
+            target_fps: None,
         }
     }
 }
@@ -82,10 +94,11 @@ impl Default for Config {
         Self {
             title: "Graphiclity Window".to_string(),
             resizeable: true,
-            logical_width: 800,
-            logical_height: 600,
-            window_width: 800,
-            window_height: 600,
+            logical_width: 640,
+            logical_height: 400,
+            window_width: 1280,
+            window_height: 800,
+            target_fps: Some(60),
         }
     }
 }